requirements.txt
flask
tensorflow
numpy
pandas
matplotlib
scikit-learn
Pillow
dill
-e .